+++ lib/red_black_tree/node/leaf_node_comparable.rb
@@ -1,0 +1,13 @@
+# frozen_string_literal: true
+
+class RedBlackTree
+  module LeafNodeComparable # @private
+    def <=> other
+      if other.instance_of? ::RedBlackTree::LeafNode
+        1
+      else
+        super
+      end
+    end
+  end
+end