+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## [Unreleased]
 
+- Alias `RedBlackTree#left_most_node` as `RedBlackTree#min`
+
 ## [0.1.1] - 2024-08-04
 
 - Update `UPDATE_WITH_YOUR_GEM_NAME_IMMEDIATELY_AFTER_RELEASE_TO_RUBYGEMS_ORG` in README